Refactor ModelValidator and make it a public provider service Separate the validation rules between core and provider-specific Use warnings as errors to enable exceptions to be thrown on validation warnings
Fixes #2416
