+++ src/Microsoft.EntityFrameworkCore.Relational/Microsoft.EntityFrameworkCore.Relational.csproj
@@ -78,7 +78,7 @@
     <Compile Include="Internal\RelationalDiagnostics.cs" />
     <Compile Include="Internal\RelationalDiagnosticSourceAfterMessage.cs" />
     <Compile Include="Internal\RelationalDiagnosticSourceBeforeMessage.cs" />
-    <Compile Include="Internal\RelationalModelValidator.cs" />
+    <Compile Include="Infrastructure\RelationalModelValidator.cs" />
     <Compile Include="Metadata\Builders\DiscriminatorBuilder.cs" />
     <Compile Include="Metadata\Builders\DiscriminatorBuilder`.cs" />
     <Compile Include="Metadata\Conventions\Internal\DiscriminatorConvention.cs" />
