+++ src/Microsoft.EntityFrameworkCore/Infrastructure/EntityFrameworkServiceCollectionExtensions.cs
@@ -90,7 +90,8 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
                 .AddSingleton<IModelCustomizer, ModelCustomizer>()
                 .AddSingleton<IModelCacheKeyFactory, ModelCacheKeyFactory>()
                 .AddSingleton<ILoggerFactory, LoggerFactory>()
-                .AddScoped<LoggingModelValidator>()
+                .AddSingleton<NoopModelValidator>()
+                .AddScoped<CoreModelValidator>()
                 .AddScoped<IKeyPropagator, KeyPropagator>()
                 .AddScoped<INavigationFixer, NavigationFixer>()
                 .AddScoped<ILocalViewListener, LocalViewListener>()
