+++ src/Microsoft.EntityFrameworkCore/Infrastructure/ModelSource.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Concurrent;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Internal;
@@ -9,6 +10,7 @@ using Microsoft.EntityFrameworkCore.Metadata.Conventions;
 using Microsoft.EntityFrameworkCore.Metadata.Conventions.Internal;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Microsoft.EntityFrameworkCore.Utilities;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Microsoft.EntityFrameworkCore.Infrastructure
 {
@@ -22,7 +24,7 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
     ///         not used in application code.
     ///     </para>
     /// </summary>
-    public abstract class ModelSource : IModelSource
+    public abstract class ModelSource : IModelSource, IServiceInjectionSite
     {
         private readonly ConcurrentDictionary<object, IModel> _models = new ConcurrentDictionary<object, IModel>();
 
@@ -37,6 +39,11 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
         ///     to build the model.
         /// </summary>
         protected virtual ICoreConventionSetBuilder CoreConventionSetBuilder { get; }
+
+        /// <summary>
+        ///     Gets the <see cref="CoreModelValidator" /> that will validate the built model.
+        /// </summary>
+        protected virtual CoreModelValidator CoreModelValidator { get; private set; }
 
         /// <summary>
         ///     Gets the <see cref="IModelCustomizer" /> that will perform additional configuration of the model
@@ -54,7 +61,12 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
         ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
         ///     directly from your code. This API may change or be removed in future releases.
         /// </summary>
-        protected ModelSource([NotNull] IDbSetFinder setFinder, [NotNull] ICoreConventionSetBuilder coreConventionSetBuilder, [NotNull] IModelCustomizer modelCustomizer, [NotNull] IModelCacheKeyFactory modelCacheKeyFactory)
+        [Obsolete("Derived classes must be updated to call the new constructor with additional parameters.")]
+        protected ModelSource(
+            [NotNull] IDbSetFinder setFinder,
+            [NotNull] ICoreConventionSetBuilder coreConventionSetBuilder,
+            [NotNull] IModelCustomizer modelCustomizer,
+            [NotNull] IModelCacheKeyFactory modelCacheKeyFactory)
         {
             Check.NotNull(setFinder, nameof(setFinder));
             Check.NotNull(coreConventionSetBuilder, nameof(coreConventionSetBuilder));
@@ -65,6 +77,30 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
             CoreConventionSetBuilder = coreConventionSetBuilder;
             ModelCustomizer = modelCustomizer;
             ModelCacheKeyFactory = modelCacheKeyFactory;
+        }
+
+        /// <summary>
+        ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
+        ///     directly from your code. This API may change or be removed in future releases.
+        /// </summary>
+        protected ModelSource(
+            [NotNull] IDbSetFinder setFinder,
+            [NotNull] ICoreConventionSetBuilder coreConventionSetBuilder,
+            [NotNull] IModelCustomizer modelCustomizer,
+            [NotNull] IModelCacheKeyFactory modelCacheKeyFactory,
+            [NotNull] CoreModelValidator coreModelValidator)
+        {
+            Check.NotNull(setFinder, nameof(setFinder));
+            Check.NotNull(coreConventionSetBuilder, nameof(coreConventionSetBuilder));
+            Check.NotNull(modelCustomizer, nameof(modelCustomizer));
+            Check.NotNull(modelCacheKeyFactory, nameof(modelCacheKeyFactory));
+            Check.NotNull(coreModelValidator, nameof(coreModelValidator));
+
+            SetFinder = setFinder;
+            CoreConventionSetBuilder = coreConventionSetBuilder;
+            ModelCustomizer = modelCustomizer;
+            ModelCacheKeyFactory = modelCacheKeyFactory;
+            CoreModelValidator = coreModelValidator;
         }
 
         /// <summary>
@@ -104,17 +140,17 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
             ModelCustomizer.Customize(modelBuilder, context);
 
             internalModelBuilder.Validate();
-
+            CoreModelValidator.Validate(modelBuilder.Model);
             validator.Validate(modelBuilder.Model);
 
             return modelBuilder.Model;
         }
 
         /// <summary>
-        ///     Creates the convention set to be used for the model. Uses the <see cref="CoreConventionSetBuilder" />
+        ///     Creates the convention set to be used for the model. Only uses the <see cref="CoreConventionSetBuilder" />
         ///     if <paramref name="conventionSetBuilder" /> is null.
         /// </summary>
-        /// <param name="conventionSetBuilder"> The convention set builder to be used. </param>
+        /// <param name="conventionSetBuilder"> The provider convention set builder to be used. </param>
         /// <returns> The convention set to be used. </returns>
         protected virtual ConventionSet CreateConventionSet([CanBeNull] IConventionSetBuilder conventionSetBuilder)
         {
@@ -136,5 +172,12 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
                 modelBuilder.Entity(setInfo.ClrType);
             }
         }
+
+        /// <summary>
+        ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
+        ///     directly from your code. This API may change or be removed in future releases.
+        /// </summary>
+        void IServiceInjectionSite.InjectServices(IServiceProvider serviceProvider)
+            => CoreModelValidator = CoreModelValidator ?? serviceProvider.GetService<CoreModelValidator>();
     }
 }
