--- src/Microsoft.EntityFrameworkCore/Internal/LoggingModelValidator.cs
@@ -1,39 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.Extensions.Logging;
-
-namespace Microsoft.EntityFrameworkCore.Internal
-{
-    /// <summary>
-    ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
-    ///     directly from your code. This API may change or be removed in future releases.
-    /// </summary>
-    public class LoggingModelValidator : ModelValidator
-    {
-        /// <summary>
-        ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
-        ///     directly from your code. This API may change or be removed in future releases.
-        /// </summary>
-        // ReSharper disable once SuggestBaseTypeForParameter
-        public LoggingModelValidator([NotNull] ILogger<LoggingModelValidator> logger)
-        {
-            Logger = logger;
-        }
-
-        /// <summary>
-        ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
-        ///     directly from your code. This API may change or be removed in future releases.
-        /// </summary>
-        protected virtual ILogger Logger { get; }
-
-        /// <summary>
-        ///     This API supports the Entity Framework Core infrastructure and is not intended to be used
-        ///     directly from your code. This API may change or be removed in future releases.
-        /// </summary>
-        protected override void ShowWarning(string message)
-            => Logger.LogWarning(CoreEventId.ModelValidationWarning, () => message);
-    }
-}
