+++ test/Microsoft.EntityFrameworkCore.InMemory.FunctionalTests/TestInMemoryModelSource.cs
@@ -19,10 +19,12 @@ namespace Microsoft.EntityFrameworkCore.InMemory.FunctionalTests
         public TestInMemoryModelSource(
             Action<ModelBuilder> onModelCreating,
             IDbSetFinder setFinder,
-            ICoreConventionSetBuilder coreConventionSetBuilder)
-            : base(setFinder, coreConventionSetBuilder, new ModelCustomizer(), new ModelCacheKeyFactory())
+            ICoreConventionSetBuilder coreConventionSetBuilder,
+            CoreModelValidator coreModelValidator)
+            : base(setFinder, coreConventionSetBuilder, new ModelCustomizer(), new ModelCacheKeyFactory(), coreModelValidator)
         {
-            _testModelSource = new TestModelSource(onModelCreating, setFinder, coreConventionSetBuilder, new ModelCustomizer(), new ModelCacheKeyFactory());
+            _testModelSource = new TestModelSource(
+                onModelCreating, setFinder, coreConventionSetBuilder, new ModelCustomizer(), new ModelCacheKeyFactory(), coreModelValidator);
         }
 
         public override IModel GetModel(DbContext context, IConventionSetBuilder conventionSetBuilder, IModelValidator validator)
@@ -32,6 +34,7 @@ namespace Microsoft.EntityFrameworkCore.InMemory.FunctionalTests
             => p => new TestInMemoryModelSource(
                 onModelCreating,
                 p.GetRequiredService<IDbSetFinder>(),
-                p.GetRequiredService<ICoreConventionSetBuilder>());
+                p.GetRequiredService<ICoreConventionSetBuilder>(),
+                p.GetRequiredService<CoreModelValidator>());
     }
 }
