+++ test/Microsoft.EntityFrameworkCore.Relational.Tests/RelationalModelValidatorTest.cs
@@ -2,18 +2,19 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.ComponentModel.DataAnnotations.Schema;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Infrastructure.Tests;
 using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Conventions.Internal;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
-using Microsoft.EntityFrameworkCore.Tests.Infrastructure;
 using Microsoft.EntityFrameworkCore.Tests.TestUtilities;
 using Microsoft.Extensions.Logging;
 using Xunit;
 
 namespace Microsoft.EntityFrameworkCore.Relational.Tests
 {
-    public class RelationalModelValidatorTest : LoggingModelValidatorTest
+    public class RelationalModelValidatorTest : ModelValidatorTest
     {
         [Fact]
         public virtual void Detects_primary_key_with_default_value()
