+++ test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/DataAnnotationSqliteFixture.cs
@@ -2,12 +2,8 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.EntityFrameworkCore.Internal;
-using Microsoft.EntityFrameworkCore.Metadata;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
+using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Specification.Tests;
-using Microsoft.EntityFrameworkCore.Storage;
-using Microsoft.EntityFrameworkCore.Storage.Internal;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
@@ -28,28 +24,6 @@ namespace Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests
                 .AddSingleton(TestSqliteModelSource.GetFactory(OnModelCreating))
                 .AddSingleton<ILoggerFactory>(new TestSqlLoggerFactory())
                 .BuildServiceProvider();
-        }
-
-        public override ModelValidator ThrowingValidator
-            => new ThrowingModelValidator(
-                _serviceProvider.GetService<ILogger<RelationalModelValidator>>(),
-                new SqliteAnnotationProvider(),
-                new SqliteTypeMapper());
-
-        private class ThrowingModelValidator : RelationalModelValidator
-        {
-            public ThrowingModelValidator(
-                ILogger<RelationalModelValidator> loggerFactory,
-                IRelationalAnnotationProvider relationalExtensions,
-                IRelationalTypeMapper typeMapper)
-                : base(loggerFactory, relationalExtensions, typeMapper)
-            {
-            }
-
-            protected override void ShowWarning(string message)
-            {
-                throw new InvalidOperationException(message);
-            }
         }
 
         public override SqliteTestStore CreateTestStore()
@@ -75,7 +49,12 @@ namespace Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests
                 .UseSqlite(
                     testStore.Connection,
                     b => b.SuppressForeignKeyEnforcement())
-                .UseInternalServiceProvider(_serviceProvider);
+                .UseInternalServiceProvider(_serviceProvider)
+                .ConfigureWarnings(w =>
+                    {
+                        w.Default(WarningBehavior.Throw);
+                        w.Ignore(CoreEventId.SensitiveDataLoggingEnabledWarning);
+                    });
 
             var context = new DataAnnotationContext(optionsBuilder.Options);
             context.Database.UseTransaction(testStore.Transaction);
