+++ test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/TestSqliteModelSource.cs
@@ -19,10 +19,11 @@ namespace Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests
         public TestSqliteModelSource(
             Action<ModelBuilder> onModelCreating,
             IDbSetFinder setFinder,
-            ICoreConventionSetBuilder coreConventionSetBuilder)
-            : base(setFinder, coreConventionSetBuilder, new ModelCustomizer(), new ModelCacheKeyFactory())
+            ICoreConventionSetBuilder coreConventionSetBuilder,
+            CoreModelValidator coreModelValidator)
+            : base(setFinder, coreConventionSetBuilder, new ModelCustomizer(), new ModelCacheKeyFactory(), coreModelValidator)
         {
-            _testModelSource = new TestModelSource(onModelCreating, setFinder, coreConventionSetBuilder, new ModelCustomizer(), new ModelCacheKeyFactory());
+            _testModelSource = new TestModelSource(onModelCreating, setFinder, coreConventionSetBuilder, new ModelCustomizer(), new ModelCacheKeyFactory(), coreModelValidator);
         }
 
         public override IModel GetModel(DbContext context, IConventionSetBuilder conventionSetBuilder, IModelValidator validator) =>
@@ -32,6 +33,7 @@ namespace Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests
             p => new TestSqliteModelSource(
                 onModelCreating,
                 p.GetRequiredService<IDbSetFinder>(),
-                p.GetRequiredService<ICoreConventionSetBuilder>());
+                p.GetRequiredService<ICoreConventionSetBuilder>(),
+                p.GetRequiredService<CoreModelValidator>());
     }
 }
