+++ test/Microsoft.EntityFrameworkCore.Tests/Metadata/Internal/InternalEntityTypeBuilderTest.cs
@@ -6,6 +6,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Conventions.Internal;
@@ -1580,7 +1581,7 @@ namespace Microsoft.EntityFrameworkCore.Tests.Metadata.Internal
             }
 
             validationConvention.Apply(modelBuilder);
-            var modelValidator = new LoggingModelValidator(new Logger<LoggingModelValidator>(new LoggerFactory()));
+            var modelValidator = new CoreModelValidator(new Logger<ModelValidator>(new LoggerFactory()));
             modelValidator.Validate(modelBuilder.Metadata);
 
             Assert.Equal(expectedIgnored, ignoredEntityTypeBuilder.Metadata.FindDeclaredIgnoredMemberConfigurationSource(memberToIgnore) == ignoreConfigurationSource);
