+++ test/Microsoft.EntityFrameworkCore.Tests/Microsoft.EntityFrameworkCore.Tests.csproj
@@ -100,10 +100,8 @@
     <Compile Include="Infrastructure\AnnotatableTest.cs" />
     <Compile Include="Infrastructure\AnnotationTest.cs" />
     <Compile Include="Infrastructure\DbContextAttributeTest.cs" />
-    <Compile Include="Infrastructure\LoggingModelValidatorTest.cs" />
+    <Compile Include="Infrastructure\CoreModelValidatorTest.cs" />
     <Compile Include="Infrastructure\ModelValidatorTest.cs" />
-    <Compile Include="Infrastructure\NonThrowingModelValidatorTest.cs" />
-    <Compile Include="Infrastructure\ThrowingModelValidatorTest.cs" />
     <Compile Include="Metadata\Conventions\ConventionSetBuilderTests.cs" />
     <Compile Include="Metadata\Conventions\Internal\BackingFieldConventionTest.cs" />
     <Compile Include="Metadata\Conventions\Internal\BaseTypeDiscoveryConventionTest.cs" />
